refactor: change handler name to route proper name
